- REST service for Vendors. - Entity, DTO, Mapper, RestController, Repository, tests.

+++ src/main/java/guru/springfamework/services/CustomerServiceImpl.java
@@ -70,6 +70,9 @@ public class CustomerServiceImpl implements CustomerService {
 
   @Override
   public void deleteCustomerById(final Long id) {
+    if (!customerRepository.existsById(id)) {
+      throw new ResourceNotFoundException();
+    }
     customerRepository.deleteById(id);
   }
